Generate CSV when uploading Sheet URL (#170)
Whenever a Google Sheet Url, its content is read through
SpreadsheetService and writed into a CSV, which is saved in the
data/ directory as the other files. By doing it this way,
no further changes are required neither in upload nor import
functionality.

`SpreadsheetService.get_data` returns an Enumerator
to avoid loading the whole content in memory while generating
the CSV.